backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
kafka-python==2.3.0
pandas==2.2.2
pyarrow==16.1.0
plotly==5.24.1
kaleido==0.2.1
